Split interfaces into each crate
