+++ kv-store/Cargo.toml
@@ -5,4 +5,12 @@ authors = ["PDAO Team <hello@postech-dao.xyz>"]
 edition = "2021"
 
 [dependencies]
+anyhow = "1.0"
+async-trait = "0.1.42"
 serde = { version = "1.0", features = ["derive"] }
+tokio = { version = "1.0", features = ["full"] }
+chrono = { version = "0.4", features = ["serde"] }
+serde_json = "1.0"
+futures = "0.3"
+log = "0.4"
+simperby-common = { version = "0.0.0", path = "../common" }
