+++ network/Cargo.toml
@@ -13,3 +13,4 @@ chrono = { version = "0.4", features = ["serde"] }
 serde_json = "1.0"
 futures = "0.3"
 log = "0.4"
+simperby-common = { version = "0.0.0", path = "../common" }
